Fix all build errors and warnings for Xcode 26
- Fix private base class cast error in Vector.hpp by using allocator() accessor
- Reorder designated initializers to match struct declaration order (ISO C++)
- Add missing field designators to fix mixed initializer warnings
- Add explicit casts for Unicode code unit type comparisons in NSStringRef.mm
- Add explicit casts for bitwise operations between different enum types
- Add explicit cast for Int-to-double implicit conversion in GlyphSpan.mm
- Bump IPHONEOS_DEPLOYMENT_TARGET from 9.3 to 12.0
- Remove dead iOS 10 availability check and deprecated openURL: fallback

+++ STULabel/Internal/GlyphSpan.mm
@@ -150,7 +150,7 @@ bool GlyphSpan::copyInnerCaretOffsetsForLigatureGlyphAtIndexImpl(
   // The font does not contain caret offsets for this glyph, so we just divide up the glyph's
   // width equally.
   n = outInnerCaretOffsets.count();
-  const CGFloat d = narrow_cast<CGFloat>(glyphSpan.typographicWidth())/(n + 1);
+  const CGFloat d = narrow_cast<CGFloat>(glyphSpan.typographicWidth())/((double)n + 1);
   CGFloat i1 = 1;
   STU_DISABLE_LOOP_UNROLL
   for (CGFloat& outOffset : outInnerCaretOffsets[{0, n}]) {
